requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
tzdata==2024.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
black==24.8.0
click==8.1.7
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
ipython==8.27.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
drf-yasg==1.21.7
inflection==0.5.1
pytz==2024.2
pyyaml==6.0.2
uritemplate==4.1.1


